+++ src/data_analysis_gui/__init__.py
@@ -4,7 +4,7 @@ Author: Charles Kissell, Northeastern University
 License: MIT (see LICENSE file for details)
 """
 
-__version__ = "0.9.1-beta.1"
+__version__ = "0.9.1-beta.2"
 
 # Import core module to make it accessible
 from . import core